Llamada a los metodos

+++ src/com/ipartek/formacion/ejercicios/herencia/Canario.java
@@ -5,7 +5,7 @@ public class Canario extends Ave {
 	
 	public String sexo;
 	public int edad;
-	public float tamanio;
+	public double tamanio;
 	
 	public Canario(String sexo, int edad) {
 		super(sexo, edad);
@@ -14,14 +14,14 @@ public class Canario extends Ave {
 		
 	}
 	
-	public Canario(String sexo, int edad, float tamanio) {
+	public Canario(String sexo, int edad, double tamanio) {
 		super(sexo, edad);
 		this.sexo = sexo;
 		this.edad = edad;
 		this.tamanio = tamanio;
 	}
 	
-	public void altura(float tamanio) {
+	public void altura(double tamanio) {
 		if(tamanio>30) {
 			System.out.println("Alto");
 		}else if(tamanio>15 || tamanio<30) {
